Move InitTextContext to helpers/ dir
This function is useful to test handlers
